remove license handler from v1
